requirements.txt
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
PyYAML==6.0.2
scikit_learn==1.3.2
torch==2.2.2
torch_geometric==2.6.1
